+++ src/game/lara/lara_cheat.h
@@ -7,4 +7,5 @@ void __cdecl Lara_Cheat_EndLevel(void);
 bool __cdecl Lara_Cheat_EnterFlyMode(void);
 bool __cdecl Lara_Cheat_ExitFlyMode(void);
 void __cdecl Lara_Cheat_GetStuff(void);
+bool Lara_Cheat_OpenNearestDoor(void);
 bool Lara_Cheat_Teleport(int32_t x, int32_t y, int32_t z);