+++ src/game/objects/vars.h
@@ -5,6 +5,8 @@
 
 extern const GAME_OBJECT_ID g_EnemyObjects[];
 extern const GAME_OBJECT_ID g_FriendObjects[];
-extern const GAME_OBJECT_ID g_InvObjects[];
 extern const GAME_OBJECT_ID g_PickupObjects[];
+extern const GAME_OBJECT_ID g_DoorObjects[];
+extern const GAME_OBJECT_ID g_TrapdoorObjects[];
+extern const GAME_OBJECT_ID g_InvObjects[];
 extern const GAME_OBJECT_PAIR g_ItemToInvObjectMap[];